+++ Tests/TuistEnvKitTests/Updater/UpdaterTests.swift
@@ -38,7 +38,7 @@ final class UpdaterTests: TuistUnitTestCase {
 
     func test_update_when_there_are_no_updates() throws {
         versionsController.semverVersionsStub = ["3.2.1"]
-        versionProvider.stubbedLatestVersionResult = .success(Version("3.2.1"))
+        versionProvider.stubbedLatestVersionResult = Version("3.2.1")
 
         try subject.update()
 
@@ -48,7 +48,7 @@ final class UpdaterTests: TuistUnitTestCase {
 
     func test_update_when_there_are_updates() throws {
         versionsController.semverVersionsStub = ["3.1.1"]
-        versionProvider.stubbedLatestVersionResult = .success(Version("3.2.1"))
+        versionProvider.stubbedLatestVersionResult = Version("3.2.1")
         var installArgs: [String] = []
         installer.installStub = { version in installArgs.append(version) }
 
@@ -62,7 +62,7 @@ final class UpdaterTests: TuistUnitTestCase {
 
     func test_update_when_no_local_versions_available() throws {
         versionsController.semverVersionsStub = []
-        versionProvider.stubbedLatestVersionResult = .success(Version("3.2.1"))
+        versionProvider.stubbedLatestVersionResult = Version("3.2.1")
 
         var installArgs: [String] = []
         installer.installStub = { version in installArgs.append(version) }